backend/pyproject.toml
[tool.poetry]
name = "cloud-mining-website"
version = "0.1.0"
description = "website"
authors = ["Axel791 <HS.Kazakov@mail.ru>"]
license = "BSD-3"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
psycopg2-binary = "^2.9.6"
httpx = "^0.24.0"
loguru = "^0.7.0"
django-split-settings = "^1.2.0"
drf-yasg = "^1.21.5"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "^5.2.2"
django-filter = "^23.2"
djoser = "^2.2.0"
celery = "^5.2.7"
flower = "^1.2.0"
redis = "^4.5.5"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
faker = "^18.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
